requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pydub==0.25.1
openai==1.57.0
deepl==1.22.0
audioop-lts==0.2.1
pytz==2024.1
